services/path-predictor/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pandas==2.1.3
numpy==1.26.2
xgboost==2.0.3
scikit-learn==1.3.2
python-dotenv==1.0.0
requests==2.31.0
joblib==1.3.2
psycopg2-binary==2.9.9
mlflow==2.9.2

